refactor(auth): migrate to PBKDF2-based key derivation for encryption
This commit overhauls the authentication and key management system to prepare for at-rest data encryption. Previosly, it used "passlib" for hashing, which was suitable for password verification only and couldnt be used to generate a key for encrypting user data.

So, to solve this, a DEK/KEK (Data Encryption Key / Key Encryption Key) architecture has been implemented:

1.  A strong, random **Data Encryption Key (DEK)** is now generated via Fernet for each user profile. This key will be used to encrypt and decrypt the user's log files.

2.  We cant store the dek in plain text so a **Key Encryption Key (KEK)** is derived from the user's masterkey and a unique, per-user "salt" using PBKDF2. This process is intentionally slow to mitigate brute-force attacks.

The KEK is then used to encrypt the DEK. The user's profile now stores the `salt` and the `encrypted_dek`, so no sensitive keys are stored directly on disk. Authentication is now handled by the ability to successfully decrypt the DEK with the user-provided masterkey.

Also, the "checkPass" function signature was updated to "(key, user)"" to reflect its changes, and all calls to the function have been modified.